+++ shelf_packages/example/lib/features/chats/controllers/chats_controller.dart
@@ -33,6 +33,4 @@ class ChatsController with RouterMixin {
 
     return JsonResponse.ok(null);
   }
-
-  Future<Response> call(Request request) => router.call(request);
 }