+++ shelf_packages/example/lib/features/chats/dto/chat_create_dto.dart
@@ -8,12 +8,24 @@ enum ChatVisibility {
   public,
 }
 
+@JsonEnum(valueField: 'code')
+enum ChatType {
+  private('PV'),
+  channel('CH'),
+  group('GR');
+
+  final String code;
+
+  const ChatType(this.code);
+}
+
 @JsonSerializable(createToJson: false)
 class ChatCreateDto {
   final String title;
   final ChatVisibility visibility;
+  final ChatType type;
 
-  const ChatCreateDto({required this.title, required this.visibility});
+  const ChatCreateDto({required this.title, required this.visibility, required this.type});
 
   factory ChatCreateDto.fromJson(Map<String, dynamic> map) => _$ChatCreateDtoFromJson(map);
 }