+++ shelf_packages/example/lib/features/messages/dto/message_dto.dart
@@ -7,8 +7,9 @@ class MessageDto {
   final String chatId;
   @JsonKey(name: 'description')
   final String content;
+  final MessageDto? parent;
 
-  const MessageDto({required this.chatId, required this.content});
+  const MessageDto({required this.chatId, required this.content, required this.parent});
 
   Map<String, dynamic> toJson() => _$MessageDtoToJson(this);
 }