+++ shelf_packages/open_api_specification/lib/src/specs/base_specs.dart
@@ -1,5 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:open_api_specification/src/specs/info_specs.dart';
+import 'package:open_api_specification/src/specs/ref_or_specs.dart';
 import 'package:open_api_specification/src/specs/schema.dart';
 import 'package:open_api_specification/src/specs/security_open_api.dart';
 import 'package:open_api_specification/src/utils/specs_serialization.dart';
@@ -223,7 +224,7 @@ class ParameterOpenApi with PrettyJsonToString {
   final String? style;
   final bool? explode;
 
-  final SchemaOpenApi? schema; // Property
+  final RefOr<SchemaOpenApi>? schema; // Property
 
   // final Map<String, MediaOpenApi> content;