+++ shelf_packages/shelf_open_api_generator/example/lib/features/messages/dto/message_dto.g.dart
@@ -12,4 +12,5 @@ Map<String, dynamic> _$MessageDtoToJson(MessageDto instance) =>
     <String, dynamic>{
       'chatId': instance.chatId,
       'description': instance.content,
+      'parent': instance.parent,
     };