+++ shelf_packages/shelf_open_api_generator/lib/src/utils/json_annotation.dart
@@ -1,10 +1,12 @@
 import 'package:analyzer/dart/element/element2.dart';
 import 'package:json_annotation/json_annotation.dart';
+import 'package:shelf_open_api_generator/src/utils/utils.dart';
 import 'package:source_gen/source_gen.dart';
 
 abstract final class JsonAnnotation {
   static const _keyChecker = TypeChecker.typeNamed(JsonKey, inPackage: 'json_annotation');
   static const _valueChecker = TypeChecker.typeNamed(JsonValue, inPackage: 'json_annotation');
+  static const _enumChecker = TypeChecker.typeNamed(JsonEnum, inPackage: 'json_annotation');
 
   static String? getFieldName(Element2 element) {
     final annotation = _keyChecker.firstAnnotationOf(element);
@@ -13,11 +15,21 @@ abstract final class JsonAnnotation {
     return annotation.getField('name')?.toStringValue();
   }
 
-  static Object? getEnumValue(Element2 element) {
-    final annotation = _valueChecker.firstAnnotationOf(element);
-    if (annotation == null) return null;
-
-    final value = annotation.getField('value')!;
-    return value.toIntValue() ?? value.toStringValue()!;
+  static Object getEnumValue(EnumElement2 enumElement, FieldElement2 fieldElement) {
+    final valueAnnotation = _valueChecker.firstAnnotationOf(fieldElement);
+    if (valueAnnotation != null) {
+      final value = valueAnnotation.getField('value')!;
+      return value.toIntValue() ?? value.toStringValue()!;
+    }
+    final enumAnnotation = _enumChecker.firstAnnotationOf(enumElement);
+    if (enumAnnotation != null) {
+      final valueEnumField = enumAnnotation.getField('valueField')?.toStringValue();
+      if (valueEnumField != null) {
+        final reader = ConstantReader(fieldElement.computeConstantValue());
+        final valueReader = reader.read(valueEnumField);
+        return valueReader.literalValue!;
+      }
+    }
+    return fieldElement.requireName;
   }
 }